nrf prototype v4

+++ libraries/GCS_MAVLink/GCS_MAVLink.h
@@ -61,7 +61,7 @@ static inline bool valid_channel(mavlink_channel_t chan)
 mavlink_message_t* mavlink_get_channel_buffer(uint8_t chan);
 mavlink_status_t* mavlink_get_channel_status(uint8_t chan);
 
-void comm_send_buffer(mavlink_channel_t chan, const uint8_t *buf, uint8_t len);
+void comm_send_buffer(mavlink_channel_t chan, const uint8_t *buf, uint16_t len);   // Crazyflie nrf support
 void comm_send_buffer_old(mavlink_channel_t chan, const uint8_t *buf, uint8_t len);
 
 /// Check for available transmit space on the nominated MAVLink channel
@@ -76,7 +76,7 @@ uint16_t comm_get_txspace(mavlink_channel_t chan);
 // lock and unlock a channel, for multi-threaded mavlink send
 void comm_send_lock(mavlink_channel_t chan, uint16_t size);
 void comm_send_unlock(mavlink_channel_t chan);
-bool is_your_msg(const uint8_t *buf, uint8_t len);
+bool is_your_msg(const uint8_t *buf, uint8_t len);    // Crazyflie nrf support
 HAL_Semaphore &comm_chan_lock(mavlink_channel_t chan);
 
 #pragma GCC diagnostic pop
